messenger/requirements.txt
uvicorn==0.17.6
fastapi==0.75.2
pydantic==1.9.0
SQLAlchemy==1.4.36
psycopg2==2.9.3
python-dotenv==0.20.0
alembic==1.7.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
celery[redis]==5.0.5
aiohttp==3.8.1
datetime
requests
uvicorn[standard]
requests
asyncio
aiofiles
Pillow==9.2.0